fix(ci): unbreak Anchor Build & Lint — prebuilt anchor-cli + clippy/fmt (#2)
The 'Install Anchor CLI' step compiled anchor-cli from source, which pulls
hidapi → libudev-dev, no longer present on ubuntu-latest runner images.
Every CI run on main failed there, so fmt/clippy/anchor-build never ran.

- Install the official prebuilt v0.32.1 Linux binary instead (~2 s, no
  system deps, no 15-min compile)
- Apply cargo fmt (never enforced until now)
- Fix the 3 clippy -D warnings errors: floating doc comment, unnecessary
  cast, manual range contains

+++ programs/soladrome/src/amm.rs
@@ -69,7 +69,7 @@ pub fn advance_pool_rewards(pool: &mut AmmPool, now: i64, rate: u32, active: boo
 
 /// Whether the continuous emission window is open at `now` for the given state.
 pub fn continuous_active(state: &ProtocolState, now: i64) -> bool {
-    (crate::state::current_epoch(now) as u64) < (state.continuous_end_epoch as u64)
+    crate::state::current_epoch(now) < u64::from(state.continuous_end_epoch)
 }
 
 /// Authority-only: approve or revoke a pool's eligibility for continuous oSOLA
@@ -82,11 +82,7 @@ pub fn set_pool_rewards(ctx: Context<SetPoolRewards>, enabled: bool) -> Result<(
     let pool = &mut ctx.accounts.pool;
     advance_pool_rewards(pool, now, rate, active);
     pool.rewards_enabled = enabled;
-    msg!(
-        "Pool {} rewards_enabled = {}",
-        pool.key(),
-        enabled,
-    );
+    msg!("Pool {} rewards_enabled = {}", pool.key(), enabled,);
     Ok(())
 }
